Create interfaces from classes.
